refactor(validator): 유효성 검사 Controller 기능으로 편입
유효성 검사 이후에 Model 쪽으로 값을 주는 것이므로 Controller 의 역할이라고 판단.
