+++ src/Controller/validator/GameCommand.js
@@ -1,7 +1,7 @@
 const { Console } = require("@woowacourse/mission-utils");
-const { ORDER, ERROR, ISALLOW } = require("../constants");
+const { ERROR, COMMAND, ISALLOW } = require("../../utils/constants");
 
-class MoveSpace {
+class GameCommand {
   #input;
 
   constructor(input) {
@@ -13,14 +13,14 @@ class MoveSpace {
       if (this.isAllowOrder()) throw new Error();
       return ISALLOW.TRUE;
     } catch (e) {
-      Console.print(ERROR.MOVE_ORDER);
+      Console.print(ERROR.GAMECOMMAND);
       return ISALLOW.FALSE;
     }
   }
 
   isAllowOrder() {
-    return this.#input !== ORDER.UP && this.#input !== ORDER.DOWN;
+    return this.#input !== COMMAND.RETRY && this.#input !== COMMAND.QUIT;
   }
 }
 
-module.exports = MoveSpace;
+module.exports = GameCommand;
