+++ src/tempCodeRunnerFile.js
@@ -1,1 +1,2 @@
 
+app.play();
